Handle scheduler generating error
The BatchHandler previously created the scheduler only in its constructor.
When AlfaNotifier restarted the agent (`stop()` then `start()`),
the stopped scheduler was not recreated in `startBatchProcessing()`.
This patch modifies `startBatchProcessing()` to check if
the scheduler is `null` or `shutdown` and creates a
new instance if necessary, allowing the agent to restart correctly.

Signed-off-by: Jun Yeong Kim <junyeonggim5@gmail.com>